Corrected Tracker program (now program check id of the task that user is entering)

+++ Chapter_002/src/main/java/ru/plahotin/trackerstrategy/DeleteItemAction.java
@@ -13,8 +13,16 @@ public class DeleteItemAction implements Action {
 	 */
 	@Override
 	public void executeAction(Input input, Tracker tracker) {
-
 		String id = input.ask("Enter id of the item to delete: ");
-		tracker.deleteItem(tracker.findById(id));
+		//boolean exists = false;
+		while (!id.equals("")) {
+			//exists = tracker.idCheck(id);
+			if (!tracker.idCheck(id)) {
+				id = input.ask("No task with such id. Please try again or press Enter to return to main menu: ");
+			} else {
+					tracker.deleteItem(tracker.findById(id));
+					break;
+			}
+		}
 	}
 }
